back/cdk/requirements.txt
aws-cdk-lib==2.211.0
constructs>=10.0.0,<11.0.0

boto3~=1.39.3
pip~=25.1.1
attrs~=25.3.0
cryptography~=45.0.5
docutils~=0.19
typing_extensions~=4.15.0
python-dateutil~=2.9.0.post0
jsii~=1.113.0
importlib_resources~=6.5.2
cattrs~=25.1.1
PyYAML~=6.0.2
publication~=0.0.3
typeguard~=2.13.3
six~=1.17.0
requests-toolbelt~=1.0.0
typing_extensions~=4.15.0
importlib_resources~=6.5.2
certifi~=2025.6.15
urllib3~=2.5.0
idna~=3.8
charset-normalizer~=3.4.2
requests~=2.32.4
awscli~=1.41.3
typing_extensions~=4.15.0
importlib_resources~=6.5.2